+++ src/pcm_resample_fallback.c
@@ -37,7 +37,7 @@ pcm_resample_fallback_deinit(struct pcm_resample_state *state)
 /* resampling code blatantly ripped from ESD */
 const int16_t *
 pcm_resample_fallback_16(struct pcm_resample_state *state,
-			 uint8_t channels,
+			 unsigned channels,
 			 unsigned src_rate,
 			 const int16_t *src_buffer, size_t src_size,
 			 unsigned dest_rate,
@@ -78,7 +78,7 @@ pcm_resample_fallback_16(struct pcm_resample_state *state,
 
 const int32_t *
 pcm_resample_fallback_32(struct pcm_resample_state *state,
-			 uint8_t channels,
+			 unsigned channels,
 			 unsigned src_rate,
 			 const int32_t *src_buffer, size_t src_size,
 			 unsigned dest_rate,
